corrected reader
Former-commit-id: 4c3f3a9d6420cb09c9df948e364041013cc41293

+++ image-processing/ReadDXI11000.m
@@ -30,4 +30,4 @@ imdata  = imread(pfname);
 
 %%% IN THIS CONFIGURATION INCREASING RADIUS IS INCREASING COL NUMBER 
 %%% INCREASING ETA IS INCREASING ROW NUMBER 
-imdata  = rot90(imdata, 2); 
+imdata  = double(rot90(imdata, 2));
